rename scanners to core
